Add files via upload
This project is divided into 3 parts. 

#1: This question concerns the construction of a NumberUtils class declaration that contains a collection of useful routines for working with integers.
#2: This question concerns the construction of a simple game called ‘Cows and Bulls’ that is
implemented using the NumberUtils class of question one. The game involves guessing a mystery 4 digit number. The number contains no repeat digits and no zero. With each guess, the number of correct digits – bulls – and the number of digits that would be correct if they were in the right position – cows. For example, say the mystery number is 8657 and the player guesses 4678, there is one bull (6), and two cows (7, 8).
#3: This program (called Question 3)  demonstrates the use of inheritance by creating and outputting 3 simple objects, where the classes for the second and third objects inherit from the class for the first object. The first object is a Shape with a name and colour. The second object is a Rectangle with additional length and width. The third object is a Circle with additional radius.